Add additional script for testing my web server

+++ 0-setup_web_static.sh
@@ -17,7 +17,7 @@ sudo echo "<html>
   <body>
     Fake HTML file for testing.
   </body>
-</html>" | sudo tee /data/web_static/releases/test/index.html
+</html>" | sudo tee /data/web_static/releases/test/index.html > /dev/null
 
 # Create symbolic link
 sudo rm -rf /data/web_static/current
@@ -28,8 +28,11 @@ sudo chown -R ubuntu:ubuntu /data/
 
 # Update Nginx configuration
 #sudo sed -i '/server_name/a \\n\tlocation /hbnb_static/ {\n\t\talias /data/web_static/current/;\n\t}\n' /etc/nginx/sites-available/default
-awk '{print} /}/ && ++count == 3 {system("echo \"\\n    location /hbnb_static/ {\n        alias /data/web_static/current/;\n    }\n' /etc/nginx/sites-available/default > mwass.txt && cat mwass.txt > /etc/nginx/sites-available/default
-rm mwass.txt
+
+if ! cat /etc/nginx/sites-available/default | grep -q "location /hbnb_static/ {"; then
+    awk '{print} /}/ && ++count == 3 {system("echo \"\\n    location /hbnb_static/ {\\n        alias /data/web_static/current/;\\n    }\"")}' /etc/nginx/sites-available/default > mwass.txt && cat mwass.txt > /etc/nginx/sites-available/default
+    rm mwass.txt
+fi
 
 # Restart Nginx
 sudo service nginx restart
